refactored some components, fixed the setUserStopped behaviour
