+++ src/screens/home/views/InfoShareView.tsx
@@ -1,67 +1,20 @@
 import React, {useCallback} from 'react';
-import {Linking, TouchableOpacity, TouchableOpacityProps, Alert} from 'react-native';
-import {Box, Text, Icon, IconProps} from 'components';
+import {Linking} from 'react-native';
+import {Box, Text} from 'components';
 import {useNavigation} from '@react-navigation/native';
 import {useI18n, useRegionalI18n} from 'locale';
 import {captureException} from 'shared/log';
 import {useStorage} from 'services/StorageService';
 import {getExposedHelpMenuURL} from 'shared/RegionLogic';
-import {
-  SystemStatus,
-  useStopExposureNotificationService,
-  useStartExposureNotificationService,
-  useSystemStatus,
-} from 'services/ExposureNotificationService';
-import NativePushNotification from 'bridge/PushNotification';
 
-interface InfoShareItemProps extends TouchableOpacityProps {
-  onPress: () => void;
-  text: string;
-  icon: IconProps['name'];
-  lastItem?: boolean;
-}
-const InfoShareItem = ({onPress, text, icon, lastItem, ...touchableProps}: InfoShareItemProps) => (
-  <>
-    <TouchableOpacity activeOpacity={0.6} onPress={onPress} accessibilityRole="button" {...touchableProps}>
-      <Box
-        paddingVertical="s"
-        marginHorizontal="-m"
-        paddingHorizontal="m"
-        flexDirection="row"
-        alignContent="center"
-        justifyContent="space-between"
-        backgroundColor="infoBlockNeutralBackground"
-        borderRadius={5}
-      >
-        <Box flex={1}>
-          <Text variant="bodyText" marginVertical="s" color="overlayBodyText">
-            {text}
-          </Text>
-        </Box>
-
-        <Box alignSelf="center">
-          <Icon size={25} name={icon} />
-        </Box>
-      </Box>
-    </TouchableOpacity>
-    {!lastItem && <Box height={5} marginHorizontal="-m" backgroundColor="overlayBackground" />}
-  </>
-);
-
-const TurnOnButton = ({systemStatus, onStart, CTA}: {systemStatus: SystemStatus; onStart: () => void; CTA: string}) => {
-  if (systemStatus !== SystemStatus.Undefined && systemStatus !== SystemStatus.Unauthorized) {
-    return <InfoShareItem onPress={onStart} text={CTA} icon="icon-chevron" lastItem />;
-  }
-  return null;
-};
+import {OnOffButton} from '../components/OnOffButton';
+import {InfoShareItem} from '../components/InfoShareItem';
 
 export const InfoShareView = () => {
   const i18n = useI18n();
-  const {region, setUserStopped} = useStorage();
+  const {region} = useStorage();
   const regionalI18n = useRegionalI18n();
   const navigation = useNavigation();
-  const startExposureNotificationService = useStartExposureNotificationService();
-  const stopExposureNotificationService = useStopExposureNotificationService();
 
   const onPrivacy = useCallback(() => {
     Linking.openURL(i18n.translate('Info.PrivacyUrl')).catch(error => captureException('An error occurred', error));
@@ -79,39 +32,6 @@ export const InfoShareView = () => {
       captureException('An error occurred', error),
     );
   }, [region, regionalI18n]);
-
-  const onStop = useCallback(() => {
-    Alert.alert(
-      i18n.translate('Info.ToggleCovidAlert.Confirm.Title'),
-      i18n.translate('Info.ToggleCovidAlert.Confirm.Body'),
-      [
-        {
-          text: i18n.translate('Info.ToggleCovidAlert.Confirm.Cancel'),
-          onPress: () => {},
-          style: 'cancel',
-        },
-        {
-          text: i18n.translate('Info.ToggleCovidAlert.Confirm.Accept'),
-          onPress: async () => {
-            await stopExposureNotificationService();
-
-            NativePushNotification.presentLocalNotification({
-              alertTitle: i18n.translate('Notification.PausedMessageTitle'),
-              alertBody: i18n.translate('Notification.PausedMessageBody'),
-            });
-          },
-          style: 'default',
-        },
-      ],
-    );
-  }, [i18n, stopExposureNotificationService]);
-
-  const onStart = useCallback(async () => {
-    await startExposureNotificationService();
-    setUserStopped(false);
-  }, [setUserStopped, startExposureNotificationService]);
-
-  const [systemStatus] = useSystemStatus();
 
   return (
     <>
@@ -151,20 +71,7 @@ export const InfoShareView = () => {
           testID="changeRegion"
         />
         <InfoShareItem onPress={onLanguage} text={i18n.translate('Info.ChangeLanguage')} icon="icon-chevron" />
-        {systemStatus === SystemStatus.Active ? (
-          <InfoShareItem
-            onPress={onStop}
-            text={i18n.translate('Info.ToggleCovidAlert.TurnOff')}
-            icon="icon-chevron"
-            lastItem
-          />
-        ) : (
-          <TurnOnButton
-            systemStatus={systemStatus}
-            onStart={onStart}
-            CTA={i18n.translate('Info.ToggleCovidAlert.TurnOn')}
-          />
-        )}
+        <OnOffButton />
       </Box>
       <Box marginTop="l" marginBottom="m">
         <Text variant="settingTitle" fontWeight="normal">
